Change font from "9x15" to "*"
My machine did not have a font with that name.

This is imperfect since the selected font may not have the expected
dimensions, but hopefully it shows something on just about any system.

There might be cases where this ends up using something like Wingdings,
which would be unfortunate; but I think that's better than no text
being displayed.